+++ client/components/Windows/Settings.vue
@@ -3,7 +3,9 @@
 		<div class="header">
 			<SidebarToggle />
 		</div>
+
 		<div class="container">
+			<h1 class="title">Settings</h1>
 			<Navigation />
 
 			<form ref="settingsForm" autocomplete="off" @change="onChange" @submit.prevent>
